+++ tests/scroll.test.js
@@ -1,14 +1,13 @@
 import React from 'react';
 import { mount } from 'enzyme';
 import List from '../src';
-import Filler from '../src/Filler';
 import { spyElementPrototypes } from './utils/domHook';
 
 function genData(count) {
   return new Array(count).fill(null).map((_, id) => ({ id }));
 }
 
-describe('List', () => {
+describe('List.Scroll', () => {
   function genList(props) {
     let node = (
       <List component="ul" itemKey="id" {...props}>
@@ -23,81 +22,7 @@ describe('List', () => {
     return mount(node);
   }
 
-  describe('raw', () => {
-    it('without height', () => {
-      const wrapper = genList({ data: genData(1) });
-      expect(wrapper.find(Filler).props().offset).toBeFalsy();
-    });
-
-    it('height over itemHeight', () => {
-      const wrapper = genList({ data: genData(1), itemHeight: 1, height: 999 });
-
-      expect(wrapper.find(Filler).props().offset).toBeFalsy();
-    });
-  });
-
-  describe('virtual', () => {
-    let scrollTop = 0;
-    let mockElement;
-
-    beforeAll(() => {
-      mockElement = spyElementPrototypes(HTMLElement, {
-        offsetHeight: {
-          get: () => 20,
-        },
-        scrollHeight: {
-          get: () => 2000,
-        },
-        clientHeight: {
-          get: () => 100,
-        },
-        scrollTop: {
-          get: () => scrollTop,
-          set(_, val) {
-            scrollTop = val;
-          },
-        },
-      });
-    });
-
-    afterAll(() => {
-      mockElement.mockRestore();
-    });
-
-    it('scroll', () => {
-      // scroll to top
-      scrollTop = 0;
-      const wrapper = genList({ itemHeight: 20, height: 100, data: genData(100) });
-      expect(wrapper.find(Filler).props().height).toEqual(2000);
-      expect(wrapper.find(Filler).props().offset).toEqual(0);
-
-      // scrollTop to end
-      scrollTop = 2000 - 100;
-      wrapper.find('ul').simulate('scroll', {
-        scrollTop,
-      });
-      expect(wrapper.find(Filler).props().height).toEqual(2000);
-      expect(wrapper.find(Filler).props().offset + wrapper.find('li').length * 20).toEqual(2000);
-    });
-
-    it('render out of view', () => {
-      scrollTop = 0;
-      let data = genData(20);
-      const onSkipRender = jest.fn();
-      const wrapper = genList({ itemHeight: 20, height: 100, disabled: true, data, onSkipRender });
-
-      data = [{ id: 'beforeAll' }, ...data];
-      wrapper.setProps({ data });
-      expect(onSkipRender).not.toHaveBeenCalled();
-
-      wrapper.setProps({ disabled: false });
-      data = [...data, { id: 'afterAll' }];
-      wrapper.setProps({ data, disabled: true });
-      expect(onSkipRender).toHaveBeenCalled();
-    });
-  });
-
-  describe('scrollTo', () => {
+  describe('scrollTo number', () => {
     const listRef = React.createRef();
     const wrapper = genList({ itemHeight: 20, height: 100, data: genData(100), ref: listRef });
 
@@ -107,26 +32,20 @@ describe('List', () => {
     });
   });
 
-  describe('status switch', () => {
-    let scrollTop = 0;
-    let scrollHeight = 0;
-
-    let mockLiElement;
+  describe('scrollTo item index', () => {
     let mockElement;
+    let scrollTop = 0;
 
     beforeAll(() => {
-      mockLiElement = spyElementPrototypes(HTMLLIElement, {
-        offsetHeight: {
-          get: () => 40,
-        },
-      });
-
       mockElement = spyElementPrototypes(HTMLElement, {
-        scrollHeight: {
-          get: () => scrollHeight,
+        offsetHeight: {
+          get: () => 20,
         },
         clientHeight: {
           get: () => 100,
+        },
+        scrollHeight: {
+          get: () => 400,
         },
         scrollTop: {
           get: () => scrollTop,
@@ -139,41 +58,49 @@ describe('List', () => {
 
     afterAll(() => {
       mockElement.mockRestore();
-      mockLiElement.mockRestore();
     });
 
-    it('raw to virtual', () => {
-      let data = genData(5);
-      const wrapper = genList({ itemHeight: 20, height: 100, data });
+    function testPlots(type, props) {
+      describe(`${type} list`, () => {
+        let listRef;
 
-      scrollHeight = 200;
-      scrollTop = 40;
-      wrapper.find('ul').simulate('scroll', {
-        scrollTop,
+        beforeEach(() => {
+          listRef = React.createRef();
+          genList({ itemHeight: 20, height: 100, data: genData(20), ref: listRef, ...props });
+        });
+
+        it('top', () => {
+          listRef.current.scrollTo({ index: 10, align: 'top' });
+          expect(scrollTop).toEqual(200);
+        });
+        it('bottom', () => {
+          listRef.current.scrollTo({ index: 10, align: 'bottom' });
+          expect(scrollTop).toEqual(120);
+        });
+        describe('auto', () => {
+          it('upper of', () => {
+            scrollTop = 210;
+            listRef.current.onScroll();
+            listRef.current.scrollTo({ index: 10, align: 'auto' });
+            expect(scrollTop).toEqual(200);
+          });
+          it('lower of', () => {
+            scrollTop = 110;
+            listRef.current.onScroll();
+            listRef.current.scrollTo({ index: 10, align: 'auto' });
+            expect(scrollTop).toEqual(120);
+          });
+          it('in range', () => {
+            scrollTop = 150;
+            listRef.current.onScroll();
+            listRef.current.scrollTo({ index: 10, align: 'auto' });
+            expect(scrollTop).toEqual(150);
+          });
+        });
       });
+    }
 
-      scrollHeight = 120;
-      data = [...data, { id: 'afterAll' }];
-      wrapper.setProps({ data });
-      expect(wrapper.find('ul').instance().scrollTop < 10).toBeTruthy();
-    });
-
-    it('virtual to raw', done => {
-      let data = genData(6);
-      const wrapper = genList({ itemHeight: 20, height: 100, data });
-
-      scrollHeight = 120;
-      scrollTop = 10;
-      wrapper.find('ul').simulate('scroll', {
-        scrollTop,
-      });
-
-      scrollHeight = 200;
-      data = data.slice(0, -1);
-      wrapper.setProps({ data });
-      expect(wrapper.find('ul').instance().scrollTop > 40).toBeTruthy();
-
-      setTimeout(done, 50);
-    });
+    testPlots('virtual list');
+    testPlots('raw list', { itemHeight: null });
   });
 });
