+++ tests/util.test.js
@@ -84,6 +84,10 @@ describe('Util', () => {
         }
       });
 
+      it('both empty', () => {
+        expect(findListDiffIndex([], [], num => num)).toEqual(null);
+      });
+
       it('same list', () => {
         const list = [1, 2, 3, 4];
         expect(findListDiffIndex(list, list, num => num)).toEqual(null);
